# docs/prd/api-contracts.yaml
openapi: 3.0.3
info:
  title: Media Transcription & Editing Service API
  description: API contracts for local media processing platform
  version: 1.0.0
  contact:
    name: Media Transcription Service
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
  /upload:
    post:
      summary: Upload media file
      description: Upload a media file for processing
      tags:
        - Media Upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Media file to upload
                metadata:
                  $ref: '#/components/schemas/MediaMetadata'
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file format or size
        '413':
          description: File too large
        '500':
          description: Server processing error
...
  /transcribe:
    post:
      summary: Transcribe media file
      description: Start transcription process with speaker diarization
      tags:
        - Transcription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mediaId:
                  type: string
                  description: ID of uploaded media file
                language:
                  type: string
                  example: "en"
                  description: Language code (auto-detect if not provided)
                enableDiarization:
                  type: boolean
                  default: true
                  description: Enable speaker diarization
                maxSpeakers:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 5
              required:
                - mediaId
      responses:
        '200':
          description: Transcription started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionJobResponse'
        '400':
          description: Invalid media ID or parameters
        '404':
          description: Media file not found
  /transcribe/{jobId}:
    get:
      summary: Get transcription status
      description: Check transcription job status and retrieve results
      tags:
        - Transcription
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: Transcription job ID
      responses:
        '200':
          description: Transcription status and results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionResult'
        '404':
          description: Job not found
        '202':
          description: Job still processing
...
components:
  schemas:
    MediaMetadata:
      type: object
      properties:
        title:
          type: string
          example: "Interview with John Doe"
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        category:
          type: string
          example: "interview"
    UploadResponse:
      type: object
      properties:
        mediaId:
          type: string
          example: "media_123456"
        filename:
          type: string
          example: "interview.mp4"
        size:
          type: integer
          example: 157286400
        duration:
          type: number
          example: 1800.5
        format:
          type: string
          example: "mp4"
        status:
          type: string
          enum: ["uploaded", "processing", "ready", "error"]
...
    TranscriptionJobResponse:
      type: object
      properties:
        jobId:
          type: string
          example: "transcribe_345678"
        mediaId:
          type: string
        status:
          type: string
          enum: ["queued", "processing", "completed", "error"]
        estimatedCompletion:
          type: string
          format: date-time
    TranscriptionResult:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: ["completed", "processing", "error"]
        transcriptId:
          type: string
        transcript:
          type: object
          properties:
            text:
              type: string
            segments:
              type: array
              items:
                $ref: '#/components/schemas/TranscriptSegment'
            speakers:
              type: array
              items:
                $ref: '#/components/schemas/Speaker'
            language:
              type: string
              example: "en"
            confidence:
              type: number
              minimum: 0
              maximum: 1
              example: 0.95
        processingTime:
          type: number
          description: Processing time in seconds
    TranscriptSegment:
      type: object
      properties:
        id:
          type: string
        start:
          type: number
          description: Start time in seconds
        end:
          type: number
          description: End time in seconds
        text:
          type: string
        speaker:
          type: string
          nullable: true
        confidence:
          type: number
          minimum: 0
          maximum: 1
        words:
          type: array
          items:
            $ref: '#/components/schemas/Word'
    Word:
      type: object
      properties:
        text:
          type: string
        start:
          type: number
        end:
          type: number
        confidence:
          type: number
    Speaker:
      type: object
      properties:
        id:
          type: string
          example: "speaker_1"
        name:
          type: string
          nullable: true
          example: "John Doe"
        totalDuration:
          type: number
          description: Total speaking time in seconds
        segmentCount:
          type: integer
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for enhanced features
security:
  - ApiKeyAuth: []
  - {} # Allow unauthenticated access for local usage